+++ tests/Unit/Public/Remove-ScriptLog.Tests.ps1
@@ -1,16 +1,16 @@
-Describe 'Close-ScriptLog' {
+Describe 'Remove-ScriptLog' {
     It 'should close the specific ScriptLog object requested (not default)' {
-        Close-ScriptLog -All
+        Remove-ScriptLog -All
         $ScriptLog1 = New-ScriptLog -LogType Memory
         $ScriptLog2 = New-ScriptLog -LogType Memory
         $ScriptLog3 = New-ScriptLog -LogType Memory
         Out-ScriptLog -Message 'One' -Log $ScriptLog1
         Out-ScriptLog -Message 'One' -Log $ScriptLog2
         Out-ScriptLog -Message 'One' -Log $ScriptLog3
-        Out-ScriptLog -Message 'Two' -Log $ScriptLog1
-        Out-ScriptLog -Message 'Three' -Log $ScriptLog1
-        Close-ScriptLog -Log $ScriptLog2
-        (Get-ScriptLog).Count | Should -Be 0
+        Out-ScriptLog -Message 'Two' -Log $ScriptLog2
+        Out-ScriptLog -Message 'Three' -Log $ScriptLog2
+        Remove-ScriptLog -Log $ScriptLog2
+        (Get-ScriptLog).Count | Should -Be 2
     }
 
 
@@ -23,8 +23,8 @@ Describe 'Close-ScriptLog' {
         Out-ScriptLog -Message 'One' -Log $ScriptLog3
         Out-ScriptLog -Message 'Two' -Log $ScriptLog1
         Out-ScriptLog -Message 'Three' -Log $ScriptLog1
-        Close-ScriptLog -All
+        Remove-ScriptLog -All
         (Get-ScriptLog).Count | Should -Be 0
     }
-    #TODO: Improve quality of tests and add additional tests to Close-ScriptLog
+    #TODO: Improve quality of tests and add additional tests to Remove-ScriptLog
 }